+++ app/client/event_listeners/listener_filter_dark.js
@@ -4,19 +4,27 @@
 // │                                     │
 // └─────────────────────────────────────┘
 
+import { remove_all_filters } from '../colour_filters/remove_all_filters.js'
+
 export function filter_dark(){
 
     const toggleFilterDark = document.getElementById('filter-dark-toggle');
 
     toggleFilterDark.addEventListener('click', async function(event) { 
 
-        if (this.dataset.toggle)
+        if (! this.dataset.toggle)
         {
-            document.querySelector('canvas').style.filter="invert(0%) grayscale(0%)";
-            this.dataset.toggle = ""
-        } else {
-            document.querySelector('canvas').style.filter="invert(90%) grayscale(90%)";
+            remove_all_filters()
+            let colourizeOne = new ol.filter.Colorize()
+            let colourizeTwo = new ol.filter.Colorize()
+            openStreetMap.addFilter(colourizeOne);
+            openStreetMap.addFilter(colourizeTwo);
+            colourizeOne.setFilter('greyscale')
+            colourizeTwo.setFilter('invert')
             this.dataset.toggle = "on"
+        } else {
+            remove_all_filters()
+            this.dataset.toggle = ""
         }
 
     });