+++ app/client/event_listeners/listener_filter_invert.js
@@ -4,18 +4,23 @@
 // │                                     │
 // └─────────────────────────────────────┘
 
+import { remove_all_filters } from '../colour_filters/remove_all_filters.js'
+
 export function filter_invert(){
     const toggleFilterInvert = document.getElementById('filter-invert-toggle');
 
     toggleFilterInvert.addEventListener('click', async function(event) { 
 
-        if (this.dataset.toggle)
+        if (! this.dataset.toggle)
         {
-            document.querySelector('canvas').style.filter="invert(0%)";
-            this.dataset.toggle = ""
-        } else {
-            document.querySelector('canvas').style.filter="invert(90%)";
+            remove_all_filters()
+            let colourize = new ol.filter.Colorize()
+            openStreetMap.addFilter(colourize);
+            colourize.setFilter('invert')
             this.dataset.toggle = "on"
+        } else {
+            remove_all_filters()
+            this.dataset.toggle = ""
         }
 
     });