+++ app/client/event_listeners/listener_refresh_page.js
@@ -1,0 +1,18 @@
+
+// ┌─────────────────────────────────────┐
+// │                                     │
+// │            Refresh Page             │
+// │                                     │
+// └─────────────────────────────────────┘
+
+export function listener_refresh_page(){
+
+    const refreshPage = document.getElementById("refresh-page")
+
+    refreshPage.addEventListener('click', async function(event) { 
+
+        // Open in chrome.
+        location.reload();
+
+    });
+}